requirements.txt
fastapi
uvicorn
pydantic
psycopg2
pymongo
bcrypt
python-jose[cryptography]
